Changed name + result will be in float

+++ miles_into_km.py
@@ -4,4 +4,4 @@ miles = int(miles)
 
 km = (miles * 1.60934)
 
-print("{} miles equals {} kilometers".format(miles, int(km)))
+print("{} miles equals {} kilometers".format(miles, float(km)))
